requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-google==10.3.0
soda-core-bigquery==3.5.1
astronomer-cosmos[dbt-bigquery]==1.9.2
protobuf==4.25.0